fix: align question subject identifier to long

+++ EduOS.Core/Entities/Exams/Question.cs
@@ -5,7 +5,7 @@ namespace EduOS.Core.Entities.Exams
 {
     public class Question : BaseTenantEntity
     {
-        public int SubjectId { get; set; }
+        public long SubjectId { get; set; }
         public int? ChapterId { get; set; }
         public string Type { get; set; } = "MCQ"; // MCQ/Short/Broad
         public string QuestionText { get; set; } = string.Empty;
